# Use bytes.Equal instead of hash in CombineTraces