**What this PR does**:
Use bytes.Equal instead of fnv hash to compare inputs in CombineTraces.  Benchmarks show a modest improvement when inputs differ (5-10%), and an extremely large improvement when inputs are identical.

Before:
```
BenchmarkCombineTraces-12             	   13642	     98505 ns/op
BenchmarkCombineTracesIdentical-12    	  184737	      6696 ns/op
```

After
```
BenchmarkCombineTraces-12             	   14634	     91460 ns/op
BenchmarkCombineTracesIdentical-12    	16753951	        76.9 ns/op
```

**Which issue(s) this PR fixes**:
n/a

**Checklist**
- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`